requirements.txt
beautifulsoup4>=4.8.2
keyboard>=0.13.5
lxml>=4.5.1
psutil>=5.7.0
pywin32>=227
pyyaml>=5.3.1
requests>=2.23.0,<3.0